fix: dtype mismatch float32/bfloat16 in forward_lm
- init_latents: accept dtype kwarg; default to model parameter dtype
  so floating_thought, last_action, qualia etc. match model precision
- forward_lm: derive dtype = next(self.parameters()).dtype once and
  pass it to init_latents + amygdala reward + NT zero tensors
- language.py: cast thought to h.dtype before from_sem projection
  (belt-and-suspenders guard for callers that don't use init_latents)

+++ neuroslm/modules/language.py
@@ -220,7 +220,7 @@ class LanguageCortex(nn.Module):
         """
         h = self.tok_emb(ids)
         if thought is not None:
-            bias = self.from_sem(thought).unsqueeze(1)  # (B, 1, d_hidden)
+            bias = self.from_sem(thought.to(h.dtype)).unsqueeze(1)  # (B, 1, d_hidden)
             h = h + bias
 
         # Compute predictive coding loss incrementally to avoid storing
